refactored client(s) into separate files
