+++ mage_ai/cenabast/pipelines/product_data_pipeline/metadata.yaml
@@ -42,7 +42,7 @@ blocks:
   language: python
   name: export_product_data
   retry_config: null
-  status: executed
+  status: updated
   timeout: null
   type: data_exporter
   upstream_blocks:
